# perf: inline `from_iter` for `ScalarBuffer`